made a simple quiz to speed up filtering. All done
I now have two lists which are roughly pure/impure but there are issues

the lists should only be of functions..I have a couple of non-function
things in there. I need to fix that

I think it makes sense to split clos off entirely as making an instance
of something still can cause side effects.

Same goes for IO and maybe streams, it's a bit hazy

Also what do we do with macros/special forms

The real answer is I need to establish what this library is for, what
does winning look like? And then I can talk about this more correctly.